refactor into lib
